dspy_rlm/pyproject.toml
[project]
name = "dspy-rlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix-otel>=0.15.0",
    "datasets>=4.6.1",
    "dspy-ai>=3.1.3",
    "fire>=0.7.1",
    "openinference-instrumentation-dspy>=0.1.33",
    "pandas>=2.3.3",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
